+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### Features
 
+- **commands/down.ts:** lvl down command added ([5f57da2](https://github.com/badassery/laravel-up/commit/5f57da2)), closes [#2](https://github.com/badassery/laravel-up/issues/2)
 - **scripts:** configure commitizen and conventional commits ([b8f47b6](https://github.com/badassery/laravel-up/commit/b8f47b6)), closes [#3](https://github.com/badassery/laravel-up/issues/3)